chore:#227 가구 컨트롤러 분리
